Map Generation Expanded
The map will now generate a set amount of rooms and then pick one room to replace with a new start point for more rooms to spawn from. This ensures that there is only one connection between each set of rooms.
Added LockTrigger to the GateManager which sense whether the player is closeby outside of the room. Later a check for a specific key will be added and allow the player to unlock this room